+++ child-learning-app/src/components/GradesView.jsx
@@ -16,7 +16,7 @@ import { LABELS, TOAST } from '../utils/messages'
 import EmptyState from './EmptyState'
 import { useFirestoreQuery } from '../hooks/useFirestoreQuery'
 import { uploadPDFToDrive, checkDriveAccess } from '../utils/googleDriveStorage'
-import { refreshGoogleAccessToken } from './Auth'
+import { refreshGoogleAccessToken } from '../utils/googleAccessToken'
 import { MAX_FILE_SIZE } from '../utils/constants'
 import { extractScoresFromImage, getGeminiUsage } from '../utils/scoreOcr'
 import DriveFilePicker from './DriveFilePicker'
