+++ child-learning-app/src/components/PastPaperView.jsx
@@ -19,7 +19,7 @@ import EmptyState from './EmptyState'
 import { toast } from '../utils/toast'
 import { LABELS, TOAST } from '../utils/messages'
 import { uploadPDFToDrive, checkDriveAccess } from '../utils/googleDriveStorage'
-import { refreshGoogleAccessToken } from './Auth'
+import { refreshGoogleAccessToken } from '../utils/googleAccessToken'
 import DriveFilePicker from './DriveFilePicker'
 import UnitTagPicker from './UnitTagPicker'
 import ProblemClipList from './ProblemClipList'
