+++ child-learning-app/src/components/PdfCropper.jsx
@@ -3,7 +3,7 @@ import { createPortal } from 'react-dom'
 import * as pdfjsLib from 'pdfjs-dist'
 import { ref, uploadBytes, getDownloadURL } from 'firebase/storage'
 import { storage } from '../firebase'
-import { getGoogleAccessToken, refreshGoogleAccessToken } from './Auth'
+import { getGoogleAccessToken, refreshGoogleAccessToken } from '../utils/googleAccessToken'
 import { getAllPDFs } from '../utils/pdfStorage'
 import Loading from './Loading'
 import './PdfCropper.css'
