+++ child-learning-app/src/components/SapixTextView.jsx
@@ -4,7 +4,7 @@ import { grades } from '../utils/unitsDatabase'
 import { subjectColors, subjectEmojis, MAX_FILE_SIZE, SUBJECTS } from '../utils/constants'
 import { getSapixTexts, addSapixText, updateSapixText, deleteSapixText } from '../utils/sapixTexts'
 import { uploadPDFToDrive, checkDriveAccess } from '../utils/googleDriveStorage'
-import { refreshGoogleAccessToken } from './Auth'
+import { refreshGoogleAccessToken } from '../utils/googleAccessToken'
 import { toast } from '../utils/toast'
 import { LABELS, TOAST } from '../utils/messages'
 import DriveFilePicker from './DriveFilePicker'
