+++ child-learning-app/src/components/TestScoreView.jsx
@@ -23,7 +23,7 @@ import ProblemClipList from './ProblemClipList'
 import PdfCropper from './PdfCropper'
 import DriveFilePicker from './DriveFilePicker'
 import { uploadPDFToDrive, checkDriveAccess } from '../utils/googleDriveStorage'
-import { refreshGoogleAccessToken } from './Auth'
+import { refreshGoogleAccessToken } from '../utils/googleAccessToken'
 import { grades } from '../utils/unitsDatabase'
 import EmptyState from './EmptyState'
 import TestRangeProblems from './TestRangeProblems'
