+++ child-learning-app/src/hooks/useFirestoreQuery.js
@@ -11,7 +11,6 @@ export function useFirestoreQuery(queryFn, deps = []) {
   const [loading, setLoading] = useState(true)
   const [error, setError] = useState(null)
 
-  // eslint-disable-next-line react-hooks/exhaustive-deps
   const load = useCallback(async () => {
     setLoading(true)
     setError(null)
